test capability equality using subset function & move utils to pkg/utils/gardener/cloudprofile.go
