+++ pkg/utils/gardener/cloudprofile.go
@@ -8,10 +8,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"maps"
 	"slices"
 
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/validation/field"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -411,4 +413,118 @@ func NewV1beta1ImagesContext(parentImages []gardencorev1beta1.MachineImage) *Ima
 			return utils.CreateMapFromSlice(mi.Versions, func(v gardencorev1beta1.MachineImageVersion) string { return v.Version })
 		},
 	)
+}
+
+// ValidateCapabilities validates the capabilities of a machine type or machine image against the capabilitiesDefinition
+// located in a cloud profile at spec.capabilities.
+// It checks if the capabilities are supported by the cloud profile and if the architecture is defined correctly.
+// It returns a list of field errors if any validation fails.
+func ValidateCapabilities(capabilities core.Capabilities, capabilitiesDefinitions []core.CapabilityDefinition, fldPath *field.Path) field.ErrorList {
+	allErrs := field.ErrorList{}
+
+	// create map from capabilitiesDefinitions
+	capabilitiesDefinition := make(map[string][]string)
+	for _, capabilityDefinition := range capabilitiesDefinitions {
+		capabilitiesDefinition[capabilityDefinition.Name] = capabilityDefinition.Values
+	}
+	supportedCapabilityKeys := slices.Collect(maps.Keys(capabilitiesDefinition))
+
+	// Check if all capabilities are supported by the cloud profile
+	for capabilityKey, capability := range capabilities {
+		supportedValues, keyExists := capabilitiesDefinition[capabilityKey]
+		if !keyExists {
+			allErrs = append(allErrs, field.NotSupported(fldPath, capabilityKey, supportedCapabilityKeys))
+			continue
+		}
+		for i, value := range capability {
+			if !slices.Contains(supportedValues, value) {
+				allErrs = append(allErrs, field.NotSupported(fldPath.Child(capabilityKey).Index(i), value, supportedValues))
+			}
+		}
+	}
+
+	// Check additional requirements for architecture
+	// must be defined when multiple architectures are supported by the cloud profile
+	supportedArchitectures := capabilitiesDefinition[v1beta1constants.ArchitectureName]
+	architectures := capabilities[v1beta1constants.ArchitectureName]
+	if len(supportedArchitectures) > 1 && len(architectures) != 1 {
+		allErrs = append(allErrs, field.Invalid(fldPath.Child(v1beta1constants.ArchitectureName), architectures, "must define exactly one architecture when multiple architectures are supported by the cloud profile"))
+	}
+
+	return allErrs
+}
+
+// GetVersionCapabilitySets returns the CapabilitySets for a given machine image version or adds the default capabilitySet if none is defined.
+func GetVersionCapabilitySets(version core.MachineImageVersion, capabilitiesDefinitions []core.CapabilityDefinition) []core.CapabilitySet {
+	versionCapabilitySets := version.CapabilitySets
+	if len(versionCapabilitySets) != 0 {
+		return versionCapabilitySets
+	}
+
+	// find architecture Capability in capabilitiesDefinitions
+	// the architecture capability must exist in every cloud profile
+	var supportedArchitectures []string
+	for _, capabilityDefinition := range capabilitiesDefinitions {
+		if capabilityDefinition.Name == v1beta1constants.ArchitectureName {
+			supportedArchitectures = capabilityDefinition.Values
+			break
+		}
+	}
+
+	// It is allowed not to define capabilitySets in the machine image version if there is only one architecture
+	// if so the capabilityDefinitions are used as default
+	if len(supportedArchitectures) == 1 {
+		capabilities := make(core.Capabilities)
+		versionCapabilitySets = []core.CapabilitySet{{Capabilities: ApplyDefaultCapabilities(capabilities, capabilitiesDefinitions)}}
+	}
+
+	return versionCapabilitySets
+}
+
+// AreCapabilitiesEqual checks if two capabilities are semantically equal.
+// It compares the keys and values of the capabilities maps.
+func AreCapabilitiesEqual(a, b core.Capabilities, capabilitiesDefinitions []core.CapabilityDefinition) bool {
+	defaultedA := ApplyDefaultCapabilities(a, capabilitiesDefinitions)
+	defaultedB := ApplyDefaultCapabilities(b, capabilitiesDefinitions)
+
+	// Check if all keys and values in `a` exist in `b` and vice versa
+	if !areCapabilitiesSubsetOf(defaultedA, defaultedB) {
+		return false
+	}
+	if !areCapabilitiesSubsetOf(defaultedB, defaultedA) {
+		return false
+	}
+
+	return true
+}
+
+// areCapabilitiesSubsetOf verifies if all keys and values in `source` exist in `target`.
+func areCapabilitiesSubsetOf(source, target core.Capabilities) bool {
+	for key, valuesSource := range source {
+		valuesTarget, exists := target[key]
+		if !exists {
+			return false
+		}
+		for _, value := range valuesSource {
+			if !slices.Contains(valuesTarget, value) {
+				return false
+			}
+		}
+	}
+	return true
+}
+
+// ApplyDefaultCapabilities sets the default capabilities based on a capabilitiesDefinition for a machine type or machine image.
+func ApplyDefaultCapabilities(capabilities core.Capabilities, capabilitiesDefinitions []core.CapabilityDefinition) core.Capabilities {
+	if len(capabilities) == 0 {
+		capabilities = make(core.Capabilities)
+	}
+
+	for _, def := range capabilitiesDefinitions {
+		if _, exists := capabilities[def.Name]; !exists {
+			capabilities[def.Name] = def.Values
+		}
+	}
+
+	return capabilities
 }
